requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
pydantic-settings
passlib[bcrypt]
pydantic[email]
python-jose[cryptography]
python-multipart
pip freeze
